[change] BaseDAO fecha a conexão ao ser destruída [refac] Código mais abstraído e reutilizável
